refactor: enum들 공통 모듈에서 관리하도록 리펙토링
현재 전부 확인하진 못했고 일부만 이동
