+++ apps/api-server/src/common/enums/currency-type.enum.ts
@@ -1,0 +1,7 @@
+/**
+ * 통화 유형을 나타내는 enum
+ */
+export enum CurrencyType {
+  KRW = 'KRW', // 원화
+  USD = 'USD', // 미국 달러
+}
