+++ apps/api-server/src/common/enums/sort-direction.enum.ts
@@ -1,0 +1,7 @@
+/**
+ * 정렬 방향을 나타내는 enum
+ */
+export enum SortDirection {
+  DESC = 'DESC', // 내림차순 (최신순)
+  ASC = 'ASC', // 오름차순 (오래된순)
+}
