+++ apps/api-server/src/common/enums/trade-status.enum.ts
@@ -1,0 +1,8 @@
+/**
+ * 거래 상태를 나타내는 enum
+ */
+export enum TradeStatus {
+  FOR_SALE = 'FOR_SALE', // 판매중
+  RESERVED = 'RESERVED', // 예약중
+  COMPLETED = 'COMPLETED', // 거래완료
+}
